debugging

+++ src/controllers/QuestController.java
@@ -185,8 +185,8 @@ public class QuestController {
         if (GameManagerController.getInstance().getCurrentLevel().getZombieWave().getCurrentWave() == 1) {
             Quest speedExecution = questsModel.getQuestByName("Speed Execution");
 
-            int currentTick = GameManagerController.getInstance().getCurrentLevel().getCurrentTick();
-            int timeWaveStarted = GameManagerController.getInstance().getCurrentLevel().getZombieWave().getTimeWaveStarted();
+            double currentTick = GameManagerController.getInstance().getCurrentLevel().getCurrentTick();
+            double timeWaveStarted = GameManagerController.getInstance().getCurrentLevel().getZombieWave().getTimeWaveStarted();
 
             if (currentTick - timeWaveStarted <= 30) {
                 speedExecution.setCurrentValue(speedExecution.getCurrentValue() + 1);
